idleness wave configuration (#803)
Cálculo de Próxima Wave por Ociosidad.
Calcula la fecha de la próxima wave por ociosidad comparando las proyecciones de backlogs Picking con los limits configurados. En caso de encontrar que es necesario generar una wave por ociosidad, obtiene la configuración de la misma a partir de los métodos ya implementados para el cálculo de los lower y upper bounds.

Las unidades a waver por cpt de las configuraciones (usadas para las proyecciones de backlogs de las próximas waves) se calculan intentando aplicar unas de las distribuciones calculadas por Waving/SOT ajustando la cantidad de unidades que se wavean en caso de que la wave precalculada no se encuentre dentro de los bounds definidos. Si no hay ninguna wave precalculada para aplicar entonces las unidades a wavear se calculan como el upper limit tomado en orden FEFO.

Issue: FWPM-1539

+++ src/main/java/com/mercadolibre/planning/model/api/projection/waverless/PrecalculatedWave.java
@@ -6,5 +6,5 @@ import lombok.Value;
 
 @Value
 public class PrecalculatedWave {
-  Map<Instant, Integer> unitsBySla;
+  Map<Instant, Long> unitsBySla;
 }
